Align path service defaltUriScheme Fixes #125366
